restructure mut cwenv
