+++ app-template/tests/integration.rs
@@ -27,7 +27,7 @@ fn setup(
     // You can set up Abstract with a builder.
     let client = AbstractClient::builder(Mock::new(&sender)).build()?;
     // The client supports setting balances for addresses and configuring ANS.
-    client.update_balance(&sender, coins(123, "ucosm"))?;
+    client.set_balance(&sender, coins(123, "ucosm"))?;
 
     // Build a Publisher Account
     let publisher = client.publisher_builder(namespace).build()?;
